backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
pydantic
pydantic-settings
python-dotenv

# Testing
pytest==8.0.0
pytest-cov==4.1.0
